Receive partner props from linked page

Relates #3
