+++ src/components/Partners/Partners.js
@@ -1,21 +1,9 @@
-import React, { useState } from "react";
+import React from "react";
 import * as SC from "../Search/Search.style";
 import * as Style from "../styles/Text.style";
-// import { Link } from "react-router-dom";
+import { Link } from "react-router-dom";
 
-const Partners = () => {
-  const [museumData, setMuseumData] = useState(null);
-
-  const museumDataCall = async () => {
-    await (await fetch("/.netlify/functions/getMuseumData/getMuseumData.js"))
-      .json()
-      .then(data => setMuseumData(data))
-      .catch(console.error);
-  };
-
-  React.useEffect(() => {
-    museumDataCall();
-  }, []);
+const Partners = ({ museumData }) => {
   if (!museumData) return <h1>Loading...</h1>;
   const museumRecord = museumData.records;
   return (
@@ -26,10 +14,23 @@ const Partners = () => {
       <SC.SearchStyle>
         {museumRecord.map(record => (
           <SC.ContentContainer key={record.fields.museum_id}>
-            <SC.ImgContainer>
-              <SC.ImgInContainer src={record.fields.museum_image} />
-            </SC.ImgContainer>
-            <p>{record.fields.name}</p>
+            <Link
+              to={{
+                pathname: "/museuminfo",
+                partnerProps: {
+                  museum_id: record.fields.museum_id,
+                  name: record.fields.name,
+                  image: record.fields.museum_image,
+                  description: record.fields.description,
+                  website: record.fields.website
+                }
+              }}
+            >
+              <SC.ImgContainer>
+                <SC.ImgInContainer src={record.fields.museum_image} />
+              </SC.ImgContainer>
+              <p>{record.fields.name}</p>
+            </Link>
           </SC.ContentContainer>
         ))}
       </SC.SearchStyle>
